[Feat]: 랭킹함수 유틸함수로 작성 및 team-result페이지에도 적용
