+++ src/types/CandidateType.ts
@@ -1,0 +1,9 @@
+export interface Candidate {
+  leaderName: string;
+  voteCount: number;
+}
+
+export interface teamProp {
+  teamName: string;
+  voteCount: number;
+}
